back-end/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
annotated-types==0.5.0
anyio==3.7.1
asgiref==3.7.2
async-timeout==4.0.2
attrs==23.1.0
beautifulsoup4==4.12.2
cachetools==5.3.1
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.4
Django==4.2.2
django-cors-headers==4.1.0
djangorestframework==3.14.0
djantic==0.7.0
drf-spectacular==0.26.3
fastapi==0.100.0
frozenlist==1.3.3
google-api-core==2.11.1
google-auth==2.21.0
google-cloud-core==2.3.2
google-cloud-translate==3.11.1
googleapis-common-protos==1.59.1
grpcio==1.56.0
grpcio-status==1.56.0
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.4
inflection==0.5.1
install==1.3.5
Jinja2==3.1.2
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.17.3
langcodes==3.3.0
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marvin==1.0.2
mdurl==0.1.2
multidict==6.0.4
numpy==1.25.0
openai==0.27.8
openai-function-call==0.1.1
pandas==2.0.3
proto-plus==1.22.3
protobuf==4.23.3
psycopg2-binary==2.9.6
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydantic==1.10.11
pydantic_core==2.0.1
Pygments==2.15.1
pyrsistent==0.19.3
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0
regex==2023.6.3
requests==2.31.0
rich==13.4.2
rsa==4.9
six==1.16.0
sniffio==1.3.0
soupsieve==2.4.1
sqlparse==0.4.4
starlette==0.27.0
tiktoken==0.4.0
tqdm==4.65.0
typer==0.9.0
typing_extensions==4.7.1
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.16
uvicorn==0.22.0
yarl==1.9.2